sample-apps/python-auto-instrumented-sample-app/requirements.txt
boto3==1.24.40
Flask==2.1.3
opentelemetry-distro==0.32b0
opentelemetry-exporter-otlp==1.12.0rc2
opentelemetry-api==1.12.0rc2
opentelemetry-instrumentation-flask==0.32b0
opentelemetry-sdk-extension-aws==2.0.1
opentelemetry-propagator-aws-xray==1.0.1
protobuf==3.20.1
PyYAML==6.0
requests==2.28.1
